+++ diktat-rules/src/main/kotlin/com/saveourtool/diktat/ruleset/rules/DiktatRule.kt
@@ -2,8 +2,8 @@ package com.saveourtool.diktat.ruleset.rules
 
 import com.saveourtool.diktat.api.DiktatErrorEmitter
 import com.saveourtool.diktat.api.DiktatRuleNameAware
+import com.saveourtool.diktat.api.isRuleEnabled
 import com.saveourtool.diktat.common.config.rules.RulesConfig
-import com.saveourtool.diktat.common.config.rules.isRuleEnabled
 import com.saveourtool.diktat.ruleset.utils.getFilePathSafely
 
 import io.github.oshai.kotlinlogging.KotlinLogging
