+++ diktat-rules/src/main/kotlin/com/saveourtool/diktat/ruleset/utils/AstNodeUtils.kt
@@ -13,9 +13,9 @@
 package com.saveourtool.diktat.ruleset.utils
 
 import com.saveourtool.diktat.DIKTAT
+import com.saveourtool.diktat.api.isAnnotatedWithIgnoredAnnotation
 import com.saveourtool.diktat.common.config.rules.Rule
 import com.saveourtool.diktat.common.config.rules.RulesConfig
-import com.saveourtool.diktat.common.config.rules.isAnnotatedWithIgnoredAnnotation
 import com.saveourtool.diktat.ruleset.rules.chapter1.PackageNaming
 
 import org.jetbrains.kotlin.KtNodeTypes
